fix: check file exists

+++ script/install_vscode.sh
@@ -10,9 +10,12 @@ tar -xzf code-stable-arm64-*.tar.gz -C $HOME/Apps
 rm code-stable-arm64-*.tar.gz
 
 # create a symlink to the code executable in /usr/local/bin
-sudo ln -s $HOME/Apps/VSCode-linux-arm64/bin/code /usr/local/bin
+if [ ! -f "/usr/local/bin/code" ]; then
+    sudo ln -s $HOME/Apps/VSCode-linux-arm64/bin/code /usr/local/bin
+fi
 
-# create a desktop entry for VSCode
+# create a desktop entry for VSCode (if needed)
+if [ ! -f "/usr/share/applications/code.desktop" ]; then
 cat << EOF | sudo tee /usr/share/applications/code.desktop
 [Desktop Entry]
 Name=Visual Studio Code
@@ -22,3 +25,4 @@ Icon=$HOME/Apps/VSCode-linux-arm64/resources/app/resources/linux/code.png
 Categories=Development;IDE;
 Terminal=false
 EOF
+fi
